Close a port/proxy

Closes the port associated with the proxy object

closeProxy(proxy, ...)

Arguments

proxy

An object of class "proxy". A proxy object see proxy.

...

Additonal function arguments

Value

invisible(proxy): An object of class "proxy" is invisibly returned. A proxy object see proxy. This allows for chaining from this function to other functions that take such an object as an argument. See examples for further details.

See also

Other session functions: emptyDNS, getPorts, openPort, setRetry, timeouts, waitForTraffic

Examples

## Not run: ------------------------------------
#   # A BMP server is assummed running on port 9090
#   prxy <- proxy(bmpPort = 9090L, port = 39500L)
#   # get ports with running proxies
#   prxy %>% getPorts
#   # set timeouts
#   prxy %>% timeouts(requestTimeout = 50000L, readTimeout = 50000L,
#                     connectionTimeout = 50000L)
#   # close proxy and check ports
#   prxy %>% 
#     closeProxy %>% 
#     getPorts
## ---------------------------------------------